[sparkle] Bring Pods and People_Agent onto the new panel system
The three full-app playgrounds each carried their own copy of the
conversation side-panel wiring, and Pods/People_Agent still had the
pre-rework version: citation-only P4, no sizing types, no fullscreen,
files opening in a sheet.

Extract the shared pieces into stories/conversationSidePanels.tsx (the
SidePanelView union, citation→DataSource mapping, label/sizing helpers,
the content renderer and the credit/files/menu top-bar actions) and wire
all three stories to it. Pods and People_Agent now get files and credit
usage panels, per-kind sizing (frames take focus, other files share),
fullscreen on file views, and pod files opening in a panel instead of the
document sheet.
